Add new secret

+++ scripts/secrets.js
@@ -6,12 +6,15 @@ const fs = require("fs");
 function secret(path) {
   const secret = crypto.randomBytes(20).toString('hex');
 
-  fs.writeFileSync(path, secret);
+  if (!fs.existsSync(path)) {
+    fs.writeFileSync(path, secret);
+  }
 }
 
 const secrets = [
   "./.secrets/db_root_password",
   "./.secrets/db_password",
+  "./.secrets/jwt_secret",
 ]
 
 secrets.forEach(secret);
